requirements.txt
db-dtypes==1.1.1
google-cloud-bigquery==3.9.0
google-cloud-storage==2.8.0
pandas==1.5.3
Pillow==8.4.0
pyarrow==12.0.0
Office365-REST-Python-Client